requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
alembic==1.11.1
amqp==5.1.1
annotated-types==0.5.0
anyio==3.7.1
async-timeout==4.0.2
asyncpg==0.28.0
attrs==23.1.0
billiard==4.1.0
celery==5.3.1
certifi==2023.7.22
cffi==1.15.1
cfgv==3.3.1
charset-normalizer==3.2.0
click==8.1.6
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
cryptography==41.0.3
distlib==0.3.7
et-xmlfile==1.1.0
fastapi==0.100.0
fastapi-cache2==0.2.1
filelock==3.12.2
flower==2.0.0
frozenlist==1.4.0
greenlet==2.0.2
gunicorn==21.2.0
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
humanize==4.7.0
identify==2.5.26
idna==3.4
iniconfig==2.0.0
kombu==5.3.1
Mako==1.2.4
MarkupSafe==2.1.3
multidict==6.0.4
nodeenv==1.8.0
numpy==1.25.2
openpyxl==3.1.2
packaging==23.1
pandas==2.0.3
pendulum==2.1.2
platformdirs==3.10.0
pluggy==1.2.0
pre-commit==3.3.3
prometheus-client==0.17.1
prompt-toolkit==3.0.39
pycparser==2.21
pydantic==2.1.1
pydantic_core==2.4.0
pytest==7.4.0
pytest-asyncio==0.21.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
pytzdata==2020.1
PyYAML==6.0.1
redis==4.6.0
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.19
starlette==0.27.0
tornado==6.3.3
types-pyOpenSSL==23.2.0.2
types-redis==4.6.0.3
typing_extensions==4.7.1
tzdata==2023.3
uvicorn==0.23.1
vine==5.0.0
virtualenv==20.24.2
wcwidth==0.2.6
yarl==1.9.2
